Better support for NULLs
In general, both bugs fixed here are the result of protojson encoding
and fields being omitted from the JSON response:

In the case of a NULL column, the NULL type is omitted from the types
enum becasue it's numeric value is "0", and protojson omits this due to
it being a default value.

In the case of `select null`, we get another edge case, where the
`values` itself is entirely omitted due to it being an empty seuqence of
bytes.

Fixes #72