Expose connection URL as a public property
In Prisma's driver adapter for Planetscale, we need to be able to get database name from the connection string. Currently, url is not exposed as a public property. In this PR we add such a property, `connectionUrl`  to both ``Client` and `Connection` classes. Unlike existing private `url` property, this one is more or less unchainged from what user have passed in a config.