Encode a Buffer as hexadecimal
This handles better for BLOB type columns. Without this, it basically
would have only worked for ASCII binary data since it gets
Buffer.toString()'d, but then does an extra quoting and sanitization
pass.

For binary data, which we can assume is coming from a Buffer, we want to
send as hexadecimal.

Fixes #161